share setup tasks between server and worker
now there is only one entry point to the ansible stuff which is
update_alpine.yml. this imports the server and the worker stuff so i
should never forget to run one before the other, because there was an
ordering problem where longhorn won't start if i run the worker
playbook before the update one on a fresh node. share all that stuff
in one file that includes our packages and set up for mount
propagation. node-exporter has been removed, because I anticipate that
i will be setting up prometheus inside the cluster and it will include
the node-exporter.